examples/generic/requirements.txt
boto3
python-dotenv
Flask